translate

+++ command/app-newmodule.js
@@ -1,6 +1,7 @@
 var chalk = require('chalk');
 var shell = require('shelljs');
 var fileExists = require('file-exists');
+var pathExists = require('path-exists');
 var replace = require('replace-in-file');
 var Newmodule = (function() {
 
@@ -10,18 +11,31 @@ var Newmodule = (function() {
 
     var _tempPath = process.config.variables.node_prefix+'/lib/node_modules/ciffi/tmp/';
 
+    pathExists(_tempPath).then(function(res) {
+      if(!res) {
+        shell.mkdir(_tempPath);
+      }
+    });
+
     var _tempFileJs = _tempPath+moduleName+'.js';
     var _resourceJs = process.config.variables.node_prefix+'/lib/node_modules/ciffi/resources/webpack/newmodule/module.js';
+    var _projectModules = process.env.PWD+'/dev/scripts/modules/';
     var _projectFileJs = process.env.PWD+'/dev/scripts/modules/'+moduleName+'.js';
 
     if(fileExists(_projectFileJs)) {
-      console.log(chalk.red('File già presente: '+_projectFileJs));
+      console.log(chalk.red('File already exists: '+_projectFileJs));
     }else {
-      shell.cp(_resourceJs,_tempFileJs);
-      replaceModuleName(_tempFileJs,moduleName,function() {
-        shell.cp(_tempFileJs,_projectFileJs);
-        shell.rm('-rf', _tempFileJs);
-        console.log(chalk.green('Nuovo file creato: '+_projectFileJs));
+      pathExists(_projectModules).then(function(res) {
+        if(res) {
+          shell.cp(_resourceJs,_tempFileJs);
+          replaceModuleName(_tempFileJs,moduleName,function() {
+            shell.cp(_tempFileJs,_projectFileJs);
+            shell.rm('-rf', _tempFileJs);
+            console.log(chalk.green('New file created: '+_projectFileJs));
+          });
+        }else {
+          console.log(chalk.red('Modules path not exists: '+_projectModules));
+        }
       });
     }
